MAINT: stats: refactoring to prepare for discrete distributions
