+++ scipy/stats/__init__.py
@@ -639,7 +639,7 @@ from ._fit import fit, goodness_of_fit
 from ._covariance import Covariance
 from ._sensitivity_analysis import *
 from ._survival import *
-from ._distribution_infrastructure import Mixture
+from ._probability_distribution import Mixture
 from ._new_distributions import Normal
 from ._mgc import multiscale_graphcorr
 from ._correlation import chatterjeexi
